Add TabmanFixedButtonBar for .buttonBar
